Mostly done w/Landing Project
I feel good about where its at. Wondering if I should
have made the entire document display in flex because
the background of the site is the dark blue color
and when the site is wide or full screen, it doesn't
stretch.
Maybe will mess with it later but not too shabby.